pyproject.toml
[tool.poetry]
name = "mission-manager"
version = "0.1.0"
description = "Service which manages drone flight path and communication with the main system."
authors = ["Szymon Romanek <srszymonsr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
albatros-uav = "^0.4.2"
confluent-kafka = "^2.2.0"
transitions = "^0.9.0"
geopy = "^2.4.0"
requests = "^2.31.0"
ortools = "^9.7.2996"
folium = "^0.14.0"


[tool.poetry.group.dev.dependencies]
pylama = "^8.4.1"
pytest = "^7.4.2"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
